Split submit-answer.test.ts by mode shape
